refactor(engine): move sql analysis and rewrite helpers into sql
