+++ packages/engine/src/engine.rs
@@ -21,12 +21,18 @@ use crate::plugin::manifest::parse_plugin_manifest_json;
 use crate::plugin::types::{InstalledPlugin, PluginManifest};
 use crate::schema_registry::{register_schema, register_schema_sql_statements};
 use crate::sql::{
-    bind_sql_with_state, build_delete_followup_sql, build_update_followup_sql,
-    coalesce_vtable_inserts_in_statement_list, escape_sql_string, expr_references_column_name,
-    parse_sql_statements,
+    active_version_from_mutations, active_version_from_update_validations, bind_sql_with_state,
+    build_delete_followup_sql, build_update_followup_sql, coalesce_lix_file_transaction_statements,
+    coalesce_vtable_inserts_in_statement_list, escape_sql_string,
+    extract_explicit_transaction_script_from_statements,
+    file_history_read_materialization_required_for_statements,
+    file_read_materialization_scope_for_statements, is_query_only_statements, parse_sql_statements,
     preprocess_parsed_statements_with_provider_and_detected_file_domain_changes, preprocess_sql,
-    ColumnReferenceOptions, DetectedFileDomainChange, MutationOperation, MutationRow,
-    PlaceholderState, PostprocessPlan, UpdateValidationPlan,
+    should_invalidate_installed_plugins_cache_for_sql,
+    should_invalidate_installed_plugins_cache_for_statements,
+    should_refresh_file_cache_for_statements, DetectedFileDomainChange,
+    FileReadMaterializationScope, MutationOperation, MutationRow, PlaceholderState,
+    PostprocessPlan,
 };
 use crate::state_commit_stream::{
     state_commit_stream_changes_from_mutations, StateCommitStream, StateCommitStreamBus,
@@ -44,13 +50,10 @@ use crate::version::{
     version_pointer_storage_version_id, DEFAULT_ACTIVE_VERSION_NAME, GLOBAL_VERSION_ID,
 };
 use crate::WasmRuntime;
-use crate::{LixBackend, LixError, LixTransaction, QueryResult, SqlDialect, Value};
+use crate::{LixBackend, LixError, LixTransaction, QueryResult, Value};
 use futures_util::FutureExt;
 use serde_json::Value as JsonValue;
-use sqlparser::ast::{
-    AssignmentTarget, BinaryOperator, Expr, FromTable, ObjectName, ObjectNamePart, Query, Select,
-    SetExpr, Statement, TableFactor, TableObject, TableWithJoins, Update, UpdateTableFromKind,
-};
+use sqlparser::ast::Statement;
 use std::collections::{BTreeMap, BTreeSet};
 use std::future::Future;
 use std::marker::PhantomData;
@@ -320,363 +323,6 @@ fn file_name_and_extension_from_path(path: &str) -> Option<(String, Option<Strin
     Some((name, extension))
 }
 
-#[cfg(test)]
-fn extract_explicit_transaction_script(
-    sql: &str,
-    params: &[Value],
-) -> Result<Option<Vec<Statement>>, LixError> {
-    let statements = parse_sql_statements(sql)?;
-    extract_explicit_transaction_script_from_statements(&statements, params)
-}
-
-fn extract_explicit_transaction_script_from_statements(
-    statements: &[Statement],
-    params: &[Value],
-) -> Result<Option<Vec<Statement>>, LixError> {
-    if !params.is_empty() {
-        return Ok(None);
-    }
-    if statements.len() < 2 {
-        return Ok(None);
-    }
-
-    let first_is_begin = matches!(statements.first(), Some(Statement::StartTransaction { .. }));
-    let last_is_commit = matches!(statements.last(), Some(Statement::Commit { .. }));
-    if !first_is_begin || !last_is_commit {
-        return Ok(None);
-    }
-
-    let middle = &statements[1..statements.len() - 1];
-    if middle.iter().any(|statement| {
-        matches!(
-            statement,
-            Statement::StartTransaction { .. }
-                | Statement::Commit { .. }
-                | Statement::Rollback { .. }
-        )
-    }) {
-        return Err(LixError {
-            message:
-                "nested transaction statements are not supported inside BEGIN ... COMMIT scripts"
-                    .to_string(),
-        });
-    }
-
-    Ok(Some(middle.to_vec()))
-}
-
-#[derive(Debug, Clone)]
-struct LixFileWriteRow {
-    id: String,
-    path_sql: String,
-    data_sql: String,
-}
-
-fn coalesce_lix_file_transaction_statements(
-    statements: &[Statement],
-    dialect: Option<SqlDialect>,
-) -> Option<Vec<String>> {
-    if statements.is_empty() {
-        return Some(Vec::new());
-    }
-    if !matches!(dialect, Some(SqlDialect::Sqlite)) {
-        return None;
-    }
-
-    let mut delete_ids = Vec::new();
-    let mut insert_rows = Vec::new();
-    let mut update_rows = Vec::new();
-    let mut seen_ids = BTreeSet::new();
-    let mut saw_update = false;
-
-    for statement in statements {
-        if let Some(ids) = parse_lix_file_delete_ids(statement) {
-            if saw_update {
-                return None;
-            }
-            for id in ids {
-                if !seen_ids.insert(id.clone()) {
-                    return None;
-                }
-                delete_ids.push(id);
-            }
-            continue;
-        }
-        if let Some(rows) = parse_lix_file_insert_rows(statement) {
-            if saw_update {
-                return None;
-            }
-            for row in rows {
-                if !seen_ids.insert(row.id.clone()) {
-                    return None;
-                }
-                insert_rows.push(row);
-            }
-            continue;
-        }
-        if let Some(row) = parse_lix_file_update_row(statement) {
-            if !seen_ids.insert(row.id.clone()) {
-                return None;
-            }
-            saw_update = true;
-            update_rows.push(row);
-            continue;
-        }
-        return None;
-    }
-
-    let mut rewritten = Vec::new();
-
-    if !delete_ids.is_empty() {
-        let id_list = delete_ids
-            .iter()
-            .map(|id| format!("'{}'", escape_sql_string(id)))
-            .collect::<Vec<_>>()
-            .join(", ");
-        rewritten.push(format!("DELETE FROM lix_file WHERE id IN ({id_list})"));
-    }
-
-    if !insert_rows.is_empty() {
-        let values = insert_rows
-            .iter()
-            .map(|row| {
-                format!(
-                    "('{}', {}, {})",
-                    escape_sql_string(&row.id),
-                    row.path_sql,
-                    row.data_sql
-                )
-            })
-            .collect::<Vec<_>>()
-            .join(", ");
-        rewritten.push(format!(
-            "INSERT INTO lix_file (id, path, data) VALUES {values}"
-        ));
-    }
-
-    if !update_rows.is_empty() {
-        let path_cases = update_rows
-            .iter()
-            .map(|row| {
-                format!(
-                    "WHEN '{}' THEN {}",
-                    escape_sql_string(&row.id),
-                    row.path_sql
-                )
-            })
-            .collect::<Vec<_>>()
-            .join(" ");
-        let data_cases = update_rows
-            .iter()
-            .map(|row| {
-                format!(
-                    "WHEN '{}' THEN {}",
-                    escape_sql_string(&row.id),
-                    row.data_sql
-                )
-            })
-            .collect::<Vec<_>>()
-            .join(" ");
-        let id_list = update_rows
-            .iter()
-            .map(|row| format!("'{}'", escape_sql_string(&row.id)))
-            .collect::<Vec<_>>()
-            .join(", ");
-        rewritten.push(format!(
-            "UPDATE lix_file \
-             SET path = CASE id {path_cases} ELSE path END, \
-                 data = CASE id {data_cases} ELSE data END \
-             WHERE id IN ({id_list})"
-        ));
-    }
-
-    Some(rewritten)
-}
-
-fn parse_lix_file_insert_rows(statement: &Statement) -> Option<Vec<LixFileWriteRow>> {
-    let Statement::Insert(insert) = statement else {
-        return None;
-    };
-    if !table_object_targets_table_name(&insert.table, "lix_file") {
-        return None;
-    }
-    if insert.columns.is_empty() {
-        return None;
-    }
-    if !insert.assignments.is_empty() || insert.on.is_some() || insert.returning.is_some() {
-        return None;
-    }
-    let source = insert.source.as_deref()?;
-    let SetExpr::Values(values) = source.body.as_ref() else {
-        return None;
-    };
-
-    if insert.columns.len() != 3 {
-        return None;
-    }
-
-    let mut id_index = None;
-    let mut path_index = None;
-    let mut data_index = None;
-    for (index, column) in insert.columns.iter().enumerate() {
-        if column.value.eq_ignore_ascii_case("id") {
-            if id_index.replace(index).is_some() {
-                return None;
-            }
-        } else if column.value.eq_ignore_ascii_case("path") {
-            if path_index.replace(index).is_some() {
-                return None;
-            }
-        } else if column.value.eq_ignore_ascii_case("data") {
-            if data_index.replace(index).is_some() {
-                return None;
-            }
-        } else {
-            // Coalescer only supports replay-shape inserts with exactly id/path/data.
-            // Any extra column must fall back to normal statement execution.
-            return None;
-        }
-    }
-    let id_index = id_index?;
-    let path_index = path_index?;
-    let data_index = data_index?;
-
-    let mut rows = Vec::with_capacity(values.rows.len());
-    for row in &values.rows {
-        let id = expr_as_string_literal(row.get(id_index)?)?;
-        let path_sql = row.get(path_index)?.to_string();
-        let data_sql = row.get(data_index)?.to_string();
-        rows.push(LixFileWriteRow {
-            id,
-            path_sql,
-            data_sql,
-        });
-    }
-    Some(rows)
-}
-
-fn parse_lix_file_update_row(statement: &Statement) -> Option<LixFileWriteRow> {
-    let Statement::Update(update) = statement else {
-        return None;
-    };
-    if !table_with_joins_targets_table_name(&update.table, "lix_file") {
-        return None;
-    }
-    if update.from.is_some() || update.returning.is_some() || update.limit.is_some() {
-        return None;
-    }
-
-    let mut path_sql = None;
-    let mut data_sql = None;
-    for assignment in &update.assignments {
-        let AssignmentTarget::ColumnName(target) = &assignment.target else {
-            return None;
-        };
-        let column = object_name_last_ident_value(target)?;
-        if column.eq_ignore_ascii_case("path") {
-            path_sql = Some(assignment.value.to_string());
-        } else if column.eq_ignore_ascii_case("data") {
-            data_sql = Some(assignment.value.to_string());
-        } else {
-            return None;
-        }
-    }
-
-    let id = parse_id_eq_selection(update.selection.as_ref()?)?;
-    Some(LixFileWriteRow {
-        id,
-        path_sql: path_sql?,
-        data_sql: data_sql?,
-    })
-}
-
-fn parse_lix_file_delete_ids(statement: &Statement) -> Option<Vec<String>> {
-    let Statement::Delete(delete) = statement else {
-        return None;
-    };
-    if !delete.tables.is_empty()
-        || delete.using.is_some()
-        || delete.returning.is_some()
-        || !delete.order_by.is_empty()
-        || delete.limit.is_some()
-    {
-        return None;
-    }
-
-    let from = match &delete.from {
-        FromTable::WithFromKeyword(from) | FromTable::WithoutKeyword(from) => from,
-    };
-    if from.len() != 1 || !table_with_joins_targets_table_name(&from[0], "lix_file") {
-        return None;
-    }
-
-    parse_id_selection(delete.selection.as_ref()?)
-}
-
-fn parse_id_selection(selection: &Expr) -> Option<Vec<String>> {
-    match selection {
-        Expr::InList {
-            expr,
-            list,
-            negated: false,
-        } if expr_is_column_name(expr, "id") => list.iter().map(expr_as_string_literal).collect(),
-        Expr::BinaryOp {
-            left,
-            op: BinaryOperator::Eq,
-            right,
-        } => {
-            if expr_is_column_name(left, "id") {
-                return Some(vec![expr_as_string_literal(right)?]);
-            }
-            if expr_is_column_name(right, "id") {
-                return Some(vec![expr_as_string_literal(left)?]);
-            }
-            None
-        }
-        _ => None,
-    }
-}
-
-fn parse_id_eq_selection(selection: &Expr) -> Option<String> {
-    let ids = parse_id_selection(selection)?;
-    if ids.len() == 1 {
-        return ids.into_iter().next();
-    }
-    None
-}
-
-fn expr_as_string_literal(expr: &Expr) -> Option<String> {
-    let Expr::Value(value) = expr else {
-        return None;
-    };
-    match &value.value {
-        sqlparser::ast::Value::SingleQuotedString(text)
-        | sqlparser::ast::Value::DoubleQuotedString(text)
-        | sqlparser::ast::Value::NationalStringLiteral(text)
-        | sqlparser::ast::Value::EscapedStringLiteral(text)
-        | sqlparser::ast::Value::UnicodeStringLiteral(text) => Some(text.clone()),
-        _ => None,
-    }
-}
-
-fn expr_is_column_name(expr: &Expr, name: &str) -> bool {
-    match expr {
-        Expr::Identifier(ident) => ident.value.eq_ignore_ascii_case(name),
-        Expr::CompoundIdentifier(parts) => parts
-            .last()
-            .is_some_and(|ident| ident.value.eq_ignore_ascii_case(name)),
-        Expr::Nested(inner) => expr_is_column_name(inner, name),
-        _ => false,
-    }
-}
-
-fn object_name_last_ident_value(name: &ObjectName) -> Option<&str> {
-    name.0
-        .last()
-        .and_then(ObjectNamePart::as_ident)
-        .map(|ident| ident.value.as_str())
-}
-
 fn collapse_pending_file_writes_for_transaction(
     writes: &[crate::filesystem::pending_file_writes::PendingFileWrite],
 ) -> Vec<crate::filesystem::pending_file_writes::PendingFileWrite> {
@@ -934,618 +580,6 @@ fn advance_placeholder_state_for_statement(
     Ok(())
 }
 
-#[cfg(test)]
-fn should_refresh_file_cache_for_sql(sql: &str) -> bool {
-    let Ok(statements) = parse_sql_statements(sql) else {
-        return false;
-    };
-    should_refresh_file_cache_for_statements(&statements)
-}
-
-fn should_refresh_file_cache_for_statements(statements: &[Statement]) -> bool {
-    statements
-        .iter()
-        .any(statement_targets_file_cache_refresh_table)
-}
-
-#[cfg(test)]
-fn is_query_only_sql(sql: &str) -> bool {
-    let Ok(statements) = parse_sql_statements(sql) else {
-        return false;
-    };
-    is_query_only_statements(&statements)
-}
-
-fn is_query_only_statements(statements: &[Statement]) -> bool {
-    !statements.is_empty()
-        && statements
-            .iter()
-            .all(|statement| matches!(statement, Statement::Query(_)))
-}
-
-fn statement_targets_file_cache_refresh_table(statement: &Statement) -> bool {
-    statement_targets_table_name(statement, "lix_state")
-        || statement_targets_table_name(statement, "lix_state_by_version")
-}
-
-fn should_invalidate_installed_plugins_cache_for_sql(sql: &str) -> bool {
-    let Ok(statements) = parse_sql_statements(sql) else {
-        return false;
-    };
-    should_invalidate_installed_plugins_cache_for_statements(&statements)
-}
-
-fn should_invalidate_installed_plugins_cache_for_statements(statements: &[Statement]) -> bool {
-    statements
-        .iter()
-        .any(|statement| statement_targets_table_name(statement, "lix_internal_plugin"))
-}
-
-fn statement_targets_table_name(statement: &Statement, table_name: &str) -> bool {
-    match statement {
-        Statement::Insert(insert) => table_object_targets_table_name(&insert.table, table_name),
-        Statement::Update(update) => table_with_joins_targets_table_name(&update.table, table_name),
-        Statement::Delete(delete) => {
-            let tables = match &delete.from {
-                FromTable::WithFromKeyword(tables) | FromTable::WithoutKeyword(tables) => tables,
-            };
-            tables
-                .iter()
-                .any(|table| table_with_joins_targets_table_name(table, table_name))
-        }
-        _ => false,
-    }
-}
-
-fn table_object_targets_table_name(table: &TableObject, table_name: &str) -> bool {
-    let TableObject::TableName(name) = table else {
-        return false;
-    };
-    object_name_targets_table_name(name, table_name)
-}
-
-fn table_with_joins_targets_table_name(table: &TableWithJoins, table_name: &str) -> bool {
-    let TableFactor::Table { name, .. } = &table.relation else {
-        return false;
-    };
-    object_name_targets_table_name(name, table_name)
-}
-
-fn object_name_targets_table_name(name: &ObjectName, table_name: &str) -> bool {
-    name.0
-        .last()
-        .and_then(ObjectNamePart::as_ident)
-        .map(|ident| ident.value.eq_ignore_ascii_case(table_name))
-        .unwrap_or(false)
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum FileReadMaterializationScope {
-    ActiveVersionOnly,
-    AllVersions,
-}
-
-#[cfg(test)]
-fn file_read_materialization_scope_for_sql(sql: &str) -> Option<FileReadMaterializationScope> {
-    let Ok(statements) = parse_sql_statements(sql) else {
-        return None;
-    };
-    file_read_materialization_scope_for_statements(&statements)
-}
-
-fn file_read_materialization_scope_for_statements(
-    statements: &[Statement],
-) -> Option<FileReadMaterializationScope> {
-    let mut scope = None;
-    for statement in statements {
-        let Some(statement_scope) = file_read_materialization_scope_for_statement(statement) else {
-            continue;
-        };
-        match statement_scope {
-            FileReadMaterializationScope::AllVersions => {
-                return Some(FileReadMaterializationScope::AllVersions);
-            }
-            FileReadMaterializationScope::ActiveVersionOnly => {
-                scope.get_or_insert(FileReadMaterializationScope::ActiveVersionOnly);
-            }
-        }
-    }
-    scope
-}
-
-#[cfg(test)]
-fn file_history_read_materialization_required_for_sql(sql: &str) -> bool {
-    let Ok(statements) = parse_sql_statements(sql) else {
-        return false;
-    };
-    file_history_read_materialization_required_for_statements(&statements)
-}
-
-fn file_history_read_materialization_required_for_statements(statements: &[Statement]) -> bool {
-    statements
-        .iter()
-        .any(file_history_read_materialization_required_for_statement)
-}
-
-fn file_history_read_materialization_required_for_statement(statement: &Statement) -> bool {
-    statement_reads_table_name(statement, "lix_file_history")
-}
-
-fn file_read_materialization_scope_for_statement(
-    statement: &Statement,
-) -> Option<FileReadMaterializationScope> {
-    let mentions_by_version = statement_reads_table_name(statement, "lix_file_by_version");
-    if mentions_by_version {
-        return Some(FileReadMaterializationScope::AllVersions);
-    }
-    if statement_reads_table_name(statement, "lix_file") {
-        return Some(FileReadMaterializationScope::ActiveVersionOnly);
-    }
-    None
-}
-
-fn statement_reads_table_name(statement: &Statement, table_name: &str) -> bool {
-    match statement {
-        Statement::Query(query) => query_mentions_table_name(query, table_name),
-        Statement::Insert(insert) => insert
-            .source
-            .as_deref()
-            .is_some_and(|query| query_mentions_table_name(query, table_name)),
-        Statement::Update(update) => {
-            let target_matches = table_with_joins_mentions_table_name(&update.table, table_name);
-            (target_matches && update_references_data_column(update))
-                || update.from.as_ref().is_some_and(|from| match from {
-                    UpdateTableFromKind::BeforeSet(from) | UpdateTableFromKind::AfterSet(from) => {
-                        from.iter()
-                            .any(|table| table_with_joins_mentions_table_name(table, table_name))
-                    }
-                })
-                || update
-                    .selection
-                    .as_ref()
-                    .is_some_and(|expr| expr_mentions_table_name(expr, table_name))
-                || update
-                    .assignments
-                    .iter()
-                    .any(|assignment| expr_mentions_table_name(&assignment.value, table_name))
-        }
-        Statement::Delete(delete) => {
-            delete.using.as_ref().is_some_and(|tables| {
-                tables
-                    .iter()
-                    .any(|table| table_with_joins_mentions_table_name(table, table_name))
-            }) || delete
-                .selection
-                .as_ref()
-                .is_some_and(|expr| expr_mentions_table_name(expr, table_name))
-        }
-        _ => false,
-    }
-}
-
-fn update_references_data_column(update: &Update) -> bool {
-    update
-        .selection
-        .as_ref()
-        .is_some_and(expr_references_data_column)
-        || update
-            .assignments
-            .iter()
-            .any(|assignment| expr_references_data_column(&assignment.value))
-}
-
-fn expr_references_data_column(expr: &Expr) -> bool {
-    expr_references_column_name(
-        expr,
-        "data",
-        ColumnReferenceOptions {
-            include_from_derived_subqueries: true,
-        },
-    )
-}
-
-fn query_mentions_table_name(query: &Query, table_name: &str) -> bool {
-    if query_set_expr_mentions_table_name(query.body.as_ref(), table_name) {
-        return true;
-    }
-
-    if let Some(with) = &query.with {
-        for cte in &with.cte_tables {
-            if query_mentions_table_name(&cte.query, table_name) {
-                return true;
-            }
-        }
-    }
-
-    if query
-        .order_by
-        .as_ref()
-        .is_some_and(|order_by| order_by_mentions_table_name(order_by, table_name))
-    {
-        return true;
-    }
-
-    if query
-        .limit_clause
-        .as_ref()
-        .is_some_and(|limit_clause| limit_clause_mentions_table_name(limit_clause, table_name))
-    {
-        return true;
-    }
-
-    if query
-        .fetch
-        .as_ref()
-        .and_then(|fetch| fetch.quantity.as_ref())
-        .is_some_and(|quantity| expr_mentions_table_name(quantity, table_name))
-    {
-        return true;
-    }
-
-    false
-}
-
-fn query_set_expr_mentions_table_name(expr: &SetExpr, table_name: &str) -> bool {
-    match expr {
-        SetExpr::Select(select) => select_mentions_table_name(select, table_name),
-        SetExpr::Query(query) => query_mentions_table_name(query, table_name),
-        SetExpr::SetOperation { left, right, .. } => {
-            query_set_expr_mentions_table_name(left.as_ref(), table_name)
-                || query_set_expr_mentions_table_name(right.as_ref(), table_name)
-        }
-        SetExpr::Values(values) => values
-            .rows
-            .iter()
-            .flatten()
-            .any(|expr| expr_mentions_table_name(expr, table_name)),
-        SetExpr::Insert(statement)
-        | SetExpr::Update(statement)
-        | SetExpr::Delete(statement)
-        | SetExpr::Merge(statement) => statement_reads_table_name(statement, table_name),
-        SetExpr::Table(table) => table
-            .table_name
-            .as_ref()
-            .is_some_and(|name| name.eq_ignore_ascii_case(table_name)),
-    }
-}
-
-fn select_mentions_table_name(select: &Select, table_name: &str) -> bool {
-    if select
-        .from
-        .iter()
-        .any(|table| table_with_joins_mentions_table_name(table, table_name))
-    {
-        return true;
-    }
-
-    if select
-        .projection
-        .iter()
-        .any(|item| select_item_mentions_table_name(item, table_name))
-    {
-        return true;
-    }
-
-    if select
-        .prewhere
-        .as_ref()
-        .is_some_and(|expr| expr_mentions_table_name(expr, table_name))
-    {
-        return true;
-    }
-
-    if select
-        .selection
-        .as_ref()
-        .is_some_and(|expr| expr_mentions_table_name(expr, table_name))
-    {
-        return true;
-    }
-
-    if group_by_expr_mentions_table_name(&select.group_by, table_name) {
-        return true;
-    }
-
-    if select
-        .cluster_by
-        .iter()
-        .any(|expr| expr_mentions_table_name(expr, table_name))
-    {
-        return true;
-    }
-
-    if select
-        .distribute_by
-        .iter()
-        .any(|expr| expr_mentions_table_name(expr, table_name))
-    {
-        return true;
-    }
-
-    if select
-        .sort_by
-        .iter()
-        .any(|order_by_expr| order_by_expr_mentions_table_name(order_by_expr, table_name))
-    {
-        return true;
-    }
-
-    if select
-        .having
-        .as_ref()
-        .is_some_and(|expr| expr_mentions_table_name(expr, table_name))
-    {
-        return true;
-    }
-
-    if select
-        .qualify
-        .as_ref()
-        .is_some_and(|expr| expr_mentions_table_name(expr, table_name))
-    {
-        return true;
-    }
-
-    if select.connect_by.as_ref().is_some_and(|connect_by| {
-        expr_mentions_table_name(&connect_by.condition, table_name)
-            || connect_by
-                .relationships
-                .iter()
-                .any(|expr| expr_mentions_table_name(expr, table_name))
-    }) {
-        return true;
-    }
-
-    false
-}
-
-fn table_with_joins_mentions_table_name(table: &TableWithJoins, table_name: &str) -> bool {
-    if table_factor_mentions_table_name(&table.relation, table_name) {
-        return true;
-    }
-
-    table.joins.iter().any(|join| {
-        table_factor_mentions_table_name(&join.relation, table_name)
-            || join_operator_mentions_table_name(&join.join_operator, table_name)
-    })
-}
-
-fn table_factor_mentions_table_name(table: &TableFactor, table_name: &str) -> bool {
-    match table {
-        TableFactor::Table { name, .. } => object_name_matches_table_name(name, table_name),
-        TableFactor::Derived { subquery, .. } => query_mentions_table_name(subquery, table_name),
-        TableFactor::NestedJoin {
-            table_with_joins, ..
-        } => table_with_joins_mentions_table_name(table_with_joins, table_name),
-        _ => false,
-    }
-}
-
-fn select_item_mentions_table_name(item: &sqlparser::ast::SelectItem, table_name: &str) -> bool {
-    match item {
-        sqlparser::ast::SelectItem::UnnamedExpr(expr)
-        | sqlparser::ast::SelectItem::ExprWithAlias { expr, .. } => {
-            expr_mentions_table_name(expr, table_name)
-        }
-        sqlparser::ast::SelectItem::QualifiedWildcard(
-            sqlparser::ast::SelectItemQualifiedWildcardKind::Expr(expr),
-            _,
-        ) => expr_mentions_table_name(expr, table_name),
-        _ => false,
-    }
-}
-
-fn group_by_expr_mentions_table_name(
-    group_by: &sqlparser::ast::GroupByExpr,
-    table_name: &str,
-) -> bool {
-    match group_by {
-        sqlparser::ast::GroupByExpr::All(_) => false,
-        sqlparser::ast::GroupByExpr::Expressions(expressions, _) => expressions
-            .iter()
-            .any(|expr| expr_mentions_table_name(expr, table_name)),
-    }
-}
-
-fn order_by_mentions_table_name(order_by: &sqlparser::ast::OrderBy, table_name: &str) -> bool {
-    match &order_by.kind {
-        sqlparser::ast::OrderByKind::All(_) => false,
-        sqlparser::ast::OrderByKind::Expressions(expressions) => expressions
-            .iter()
-            .any(|expr| order_by_expr_mentions_table_name(expr, table_name)),
-    }
-}
-
-fn order_by_expr_mentions_table_name(
-    order_by_expr: &sqlparser::ast::OrderByExpr,
-    table_name: &str,
-) -> bool {
-    if expr_mentions_table_name(&order_by_expr.expr, table_name) {
-        return true;
-    }
-
-    order_by_expr.with_fill.as_ref().is_some_and(|with_fill| {
-        with_fill
-            .from
-            .as_ref()
-            .is_some_and(|expr| expr_mentions_table_name(expr, table_name))
-            || with_fill
-                .to
-                .as_ref()
-                .is_some_and(|expr| expr_mentions_table_name(expr, table_name))
-            || with_fill
-                .step
-                .as_ref()
-                .is_some_and(|expr| expr_mentions_table_name(expr, table_name))
-    })
-}
-
-fn limit_clause_mentions_table_name(
-    limit_clause: &sqlparser::ast::LimitClause,
-    table_name: &str,
-) -> bool {
-    match limit_clause {
-        sqlparser::ast::LimitClause::LimitOffset {
-            limit,
-            offset,
-            limit_by,
-        } => {
-            limit
-                .as_ref()
-                .is_some_and(|expr| expr_mentions_table_name(expr, table_name))
-                || offset
-                    .as_ref()
-                    .is_some_and(|offset| expr_mentions_table_name(&offset.value, table_name))
-                || limit_by
-                    .iter()
-                    .any(|expr| expr_mentions_table_name(expr, table_name))
-        }
-        sqlparser::ast::LimitClause::OffsetCommaLimit { offset, limit } => {
-            expr_mentions_table_name(offset, table_name)
-                || expr_mentions_table_name(limit, table_name)
-        }
-    }
-}
-
-fn join_operator_mentions_table_name(
-    join_operator: &sqlparser::ast::JoinOperator,
-    table_name: &str,
-) -> bool {
-    let (match_condition, constraint) = match join_operator {
-        sqlparser::ast::JoinOperator::AsOf {
-            match_condition,
-            constraint,
-        } => (Some(match_condition), Some(constraint)),
-        sqlparser::ast::JoinOperator::Join(constraint)
-        | sqlparser::ast::JoinOperator::Inner(constraint)
-        | sqlparser::ast::JoinOperator::Left(constraint)
-        | sqlparser::ast::JoinOperator::LeftOuter(constraint)
-        | sqlparser::ast::JoinOperator::Right(constraint)
-        | sqlparser::ast::JoinOperator::RightOuter(constraint)
-        | sqlparser::ast::JoinOperator::FullOuter(constraint)
-        | sqlparser::ast::JoinOperator::CrossJoin(constraint)
-        | sqlparser::ast::JoinOperator::Semi(constraint)
-        | sqlparser::ast::JoinOperator::LeftSemi(constraint)
-        | sqlparser::ast::JoinOperator::RightSemi(constraint)
-        | sqlparser::ast::JoinOperator::Anti(constraint)
-        | sqlparser::ast::JoinOperator::LeftAnti(constraint)
-        | sqlparser::ast::JoinOperator::RightAnti(constraint)
-        | sqlparser::ast::JoinOperator::StraightJoin(constraint) => (None, Some(constraint)),
-        sqlparser::ast::JoinOperator::CrossApply | sqlparser::ast::JoinOperator::OuterApply => {
-            (None, None)
-        }
-    };
-
-    match_condition.is_some_and(|expr| expr_mentions_table_name(expr, table_name))
-        || constraint
-            .is_some_and(|constraint| join_constraint_mentions_table_name(constraint, table_name))
-}
-
-fn join_constraint_mentions_table_name(
-    constraint: &sqlparser::ast::JoinConstraint,
-    table_name: &str,
-) -> bool {
-    match constraint {
-        sqlparser::ast::JoinConstraint::On(expr) => expr_mentions_table_name(expr, table_name),
-        _ => false,
-    }
-}
-
-fn expr_mentions_table_name(expr: &Expr, table_name: &str) -> bool {
-    match expr {
-        Expr::BinaryOp { left, right, .. } => {
-            expr_mentions_table_name(left, table_name)
-                || expr_mentions_table_name(right, table_name)
-        }
-        Expr::UnaryOp { expr, .. }
-        | Expr::Nested(expr)
-        | Expr::IsNull(expr)
-        | Expr::IsNotNull(expr)
-        | Expr::Cast { expr, .. } => expr_mentions_table_name(expr, table_name),
-        Expr::InList { expr, list, .. } => {
-            expr_mentions_table_name(expr, table_name)
-                || list
-                    .iter()
-                    .any(|item| expr_mentions_table_name(item, table_name))
-        }
-        Expr::InSubquery { expr, subquery, .. } => {
-            expr_mentions_table_name(expr, table_name)
-                || query_mentions_table_name(subquery, table_name)
-        }
-        Expr::Between {
-            expr, low, high, ..
-        } => {
-            expr_mentions_table_name(expr, table_name)
-                || expr_mentions_table_name(low, table_name)
-                || expr_mentions_table_name(high, table_name)
-        }
-        Expr::Like { expr, pattern, .. } | Expr::ILike { expr, pattern, .. } => {
-            expr_mentions_table_name(expr, table_name)
-                || expr_mentions_table_name(pattern, table_name)
-        }
-        Expr::InUnnest {
-            expr, array_expr, ..
-        } => {
-            expr_mentions_table_name(expr, table_name)
-                || expr_mentions_table_name(array_expr, table_name)
-        }
-        Expr::AnyOp { left, right, .. } | Expr::AllOp { left, right, .. } => {
-            expr_mentions_table_name(left, table_name)
-                || expr_mentions_table_name(right, table_name)
-        }
-        Expr::Exists { subquery, .. } | Expr::Subquery(subquery) => {
-            query_mentions_table_name(subquery, table_name)
-        }
-        Expr::Function(function) => match &function.args {
-            sqlparser::ast::FunctionArguments::List(list) => {
-                list.args.iter().any(|arg| match arg {
-                    sqlparser::ast::FunctionArg::Unnamed(
-                        sqlparser::ast::FunctionArgExpr::Expr(expr),
-                    ) => expr_mentions_table_name(expr, table_name),
-                    sqlparser::ast::FunctionArg::Named { arg, .. }
-                    | sqlparser::ast::FunctionArg::ExprNamed { arg, .. } => match arg {
-                        sqlparser::ast::FunctionArgExpr::Expr(expr) => {
-                            expr_mentions_table_name(expr, table_name)
-                        }
-                        _ => false,
-                    },
-                    _ => false,
-                })
-            }
-            _ => false,
-        },
-        Expr::Case {
-            operand,
-            conditions,
-            else_result,
-            ..
-        } => {
-            operand
-                .as_ref()
-                .is_some_and(|operand| expr_mentions_table_name(operand, table_name))
-                || conditions.iter().any(|condition| {
-                    expr_mentions_table_name(&condition.condition, table_name)
-                        || expr_mentions_table_name(&condition.result, table_name)
-                })
-                || else_result
-                    .as_ref()
-                    .is_some_and(|value| expr_mentions_table_name(value, table_name))
-        }
-        Expr::Tuple(items) => items
-            .iter()
-            .any(|item| expr_mentions_table_name(item, table_name)),
-        _ => false,
-    }
-}
-
-fn object_name_matches_table_name(name: &ObjectName, table_name: &str) -> bool {
-    name.0
-        .last()
-        .and_then(ObjectNamePart::as_ident)
-        .map(|ident| ident.value.eq_ignore_ascii_case(table_name))
-        .unwrap_or(false)
-}
-
 fn builtin_schema_entity_id(schema: &JsonValue) -> Result<String, LixError> {
     let schema_key = schema
         .get("x-lix-key")
@@ -1563,199 +597,26 @@ fn builtin_schema_entity_id(schema: &JsonValue) -> Result<String, LixError> {
     Ok(format!("{schema_key}~{schema_version}"))
 }
 
-fn active_version_from_mutations(mutations: &[MutationRow]) -> Result<Option<String>, LixError> {
-    for mutation in mutations.iter().rev() {
-        if !mutation.untracked {
-            continue;
-        }
-        if mutation.schema_key != active_version_schema_key()
-            || mutation.file_id != active_version_file_id()
-            || mutation.version_id != active_version_storage_version_id()
-        {
-            continue;
-        }
-
-        let snapshot = mutation.snapshot_content.as_ref().ok_or_else(|| LixError {
-            message: "active version mutation is missing snapshot_content".to_string(),
-        })?;
-        let snapshot_content = serde_json::to_string(snapshot).map_err(|error| LixError {
-            message: format!("active version mutation snapshot_content invalid JSON: {error}"),
-        })?;
-        return parse_active_version_snapshot(&snapshot_content).map(Some);
-    }
-
-    Ok(None)
-}
-
-fn active_version_from_update_validations(
-    plans: &[UpdateValidationPlan],
-) -> Result<Option<String>, LixError> {
-    for plan in plans.iter().rev() {
-        if !plan
-            .table
-            .eq_ignore_ascii_case("lix_internal_state_untracked")
-        {
-            continue;
-        }
-        if !where_clause_targets_active_version(plan.where_clause.as_ref()) {
-            continue;
-        }
-        let Some(snapshot) = plan.snapshot_content.as_ref() else {
-            continue;
-        };
-
-        let snapshot_content = serde_json::to_string(snapshot).map_err(|error| LixError {
-            message: format!("active version update snapshot_content invalid JSON: {error}"),
-        })?;
-        return parse_active_version_snapshot(&snapshot_content).map(Some);
-    }
-
-    Ok(None)
-}
-
-fn where_clause_targets_active_version(where_clause: Option<&Expr>) -> bool {
-    let Some(where_clause) = where_clause else {
-        return false;
-    };
-    let Some(schema_keys) = schema_keys_from_expr(where_clause) else {
-        return false;
-    };
-    schema_keys
-        .iter()
-        .any(|value| value.eq_ignore_ascii_case(active_version_schema_key()))
-}
-
-fn schema_keys_from_expr(expr: &Expr) -> Option<Vec<String>> {
-    match expr {
-        Expr::BinaryOp {
-            left,
-            op: BinaryOperator::Eq,
-            right,
-        } => {
-            if expr_is_schema_key_column(left) {
-                return schema_key_literal_value(right).map(|value| vec![value]);
-            }
-            if expr_is_schema_key_column(right) {
-                return schema_key_literal_value(left).map(|value| vec![value]);
-            }
-            None
-        }
-        Expr::BinaryOp {
-            left,
-            op: BinaryOperator::And,
-            right,
-        } => match (schema_keys_from_expr(left), schema_keys_from_expr(right)) {
-            (Some(left), Some(right)) => {
-                let intersection = intersect_strings(&left, &right);
-                if intersection.is_empty() {
-                    None
-                } else {
-                    Some(intersection)
-                }
-            }
-            (Some(keys), None) | (None, Some(keys)) => Some(keys),
-            (None, None) => None,
-        },
-        Expr::BinaryOp {
-            left,
-            op: BinaryOperator::Or,
-            right,
-        } => match (schema_keys_from_expr(left), schema_keys_from_expr(right)) {
-            (Some(left), Some(right)) => Some(union_strings(&left, &right)),
-            _ => None,
-        },
-        Expr::InList {
-            expr,
-            list,
-            negated: false,
-        } => {
-            if !expr_is_schema_key_column(expr) {
-                return None;
-            }
-            let mut values = Vec::with_capacity(list.len());
-            for item in list {
-                let value = schema_key_literal_value(item)?;
-                values.push(value);
-            }
-            if values.is_empty() {
-                None
-            } else {
-                Some(dedup_strings(values))
-            }
-        }
-        Expr::Nested(inner) => schema_keys_from_expr(inner),
-        _ => None,
-    }
-}
-
-fn expr_is_schema_key_column(expr: &Expr) -> bool {
-    match expr {
-        Expr::Identifier(ident) => ident.value.eq_ignore_ascii_case("schema_key"),
-        Expr::CompoundIdentifier(idents) => idents
-            .last()
-            .map(|ident| ident.value.eq_ignore_ascii_case("schema_key"))
-            .unwrap_or(false),
-        _ => false,
-    }
-}
-
-fn schema_key_literal_value(expr: &Expr) -> Option<String> {
-    match expr {
-        Expr::Value(value) => value.value.clone().into_string(),
-        Expr::Identifier(ident) if ident.quote_style == Some('"') => Some(ident.value.clone()),
-        _ => None,
-    }
-}
-
-fn dedup_strings(values: Vec<String>) -> Vec<String> {
-    let mut out = Vec::new();
-    for value in values {
-        if !out.contains(&value) {
-            out.push(value);
-        }
-    }
-    out
-}
-
-fn union_strings(left: &[String], right: &[String]) -> Vec<String> {
-    let mut out = left.to_vec();
-    for value in right {
-        if !out.contains(value) {
-            out.push(value.clone());
-        }
-    }
-    out
-}
-
-fn intersect_strings(left: &[String], right: &[String]) -> Vec<String> {
-    let mut out = Vec::new();
-    for value in left {
-        if right.contains(value) && !out.contains(value) {
-            out.push(value.clone());
-        }
-    }
-    out
-}
-
 #[cfg(test)]
 mod tests {
     use super::{
-        active_version_from_update_validations, active_version_schema_key,
-        advance_placeholder_state_for_statement, boot, coalesce_lix_file_transaction_statements,
+        advance_placeholder_state_for_statement, boot,
         detected_file_domain_changes_from_detected_file_changes,
-        detected_file_domain_changes_with_writer_key, extract_explicit_transaction_script,
-        file_descriptor_cache_eviction_targets, file_history_read_materialization_required_for_sql,
-        file_read_materialization_scope_for_sql, is_query_only_sql,
-        should_invalidate_installed_plugins_cache_for_sql, should_refresh_file_cache_for_sql,
+        detected_file_domain_changes_with_writer_key, file_descriptor_cache_eviction_targets,
         should_sequentialize_postprocess_multi_statement, BootArgs, ExecuteOptions,
-        FileReadMaterializationScope,
     };
     use crate::backend::{LixBackend, LixTransaction, SqlDialect};
     use crate::plugin::types::{InstalledPlugin, PluginRuntime};
     use crate::sql::{
-        bind_sql_with_state, parse_sql_statements, MutationOperation, MutationRow, PlaceholderState,
+        active_version_from_update_validations, bind_sql_with_state,
+        coalesce_lix_file_transaction_statements, extract_explicit_transaction_script,
+        file_history_read_materialization_required_for_sql,
+        file_read_materialization_scope_for_sql, is_query_only_sql, parse_sql_statements,
+        should_invalidate_installed_plugins_cache_for_sql, should_refresh_file_cache_for_sql,
+        FileReadMaterializationScope, MutationOperation, MutationRow, PlaceholderState,
     };
     use crate::sql::{DetectedFileDomainChange, UpdateValidationPlan};
+    use crate::version::active_version_schema_key;
     use crate::{LixError, QueryResult, SnapshotChunkReader, Value, WasmComponentInstance};
     use async_trait::async_trait;
     use serde_json::json;
